Filter certificate list with name/email of current key/group
If there is already a key or group selected as recipient, then filter the
certificate list with name or email of the key resp. name of the group
because simply filtering with the text from the input field (which is
the pretty summary text for the key or group) most likely won't yield any
matching certificates or groups.

GnuPG-bug-id: 5845