Use locateKeysJob instead of keyForMailboxJob for locating key
KeyForMailboxJob internally does a KeyListJob with the same options
as the KeyListJob created by locateKeysJob() and then tries to find
the best key matching the given text. But here we are only interested
in triggering the look up and import of keys. The result of the job is
ignored. Therefore, we can use the more simple locateKeysJob().

Additionally, we remove leading and trailing whitespace and do not
perform a look up if the entered text is empty.

GnuPG-bug-id: 5845