Use the key cache instead of running our own keylist job
Running our own keylist job causes unnecessary load. Moreover, we miss
later updates of the key list. Using the key cache to look for uncertified
OpenPGP keys fixes both issues. In particular, the button is hidden/shown
automagically when the last uncertified key is certified or when the
certification of a certified key is revoked.

GnuPG-bug-id: 5850