Prevent empty lines in the revocation comment
The edit interface that is used for revoking own OpenPGP keys allows
entering a comment/description consisting of multiple lines of text.
The text ends with an empty line. Therefore, it is not possible to
add revocation comments that contain empty lines. Any whitespace at the
end of lines and any trailing whitespace at the end of the entered text
is stripped.

GnuPG-bug-id: 5859