Allow marking a form input as required
This adds functionality to require a text input field to be filled out.
If input is required, then this will be indicated to assistive tools.
A special error message, that is displayed, if a required field lacks
input, can be set. hasValue() can be used to check if a value has been
entered.

GnuPG-bug-id: 5916