Visually indicate required input fields
To indicate required input fields the text " (required)" is appended
to the label. We do not add "required" to the accessible name anymore.
In the Add User ID dialog we remove the colon at the end of the labels
texts and we do not need to set an alternative accessible name anymore.

GnuPG-bug-id: 5916