Update usage flags even if key type is forced
Separate the update of the enabled state of the widgets and the checked
state of the widgets. Update the enabled state only if the key type is
not forced, but always update the checked state.

This fixes the selection of the correct usage flags if the key type is
forced.

GnuPG-bug-id: 5865