Minor changes of our scroll area
* In the c'tor we add the layout to the widget before we set the widget as
the scroll area's widget because the API documentation of
QScrollArea::setWidget includes a note that this is required to avoid
problems with the visibility of the widget. Those problems didn't
happen for us, but we better play be the rules.
* In minimumSizeHint the height of the widget's minimum size hint is now
added.
* In sizeHint we use viewportSizeHint() (which uses the actual size of
resizable widgets) instead of the widget's size hint.
* For both size hints the width of the vertical scroll bar is not added
anymore if it is explicitly turned off. The height of the horizontal
scroll bar is added now unless it is explicitly turned off.

I have checked that the current usages of ScrollArea still work as
expected.

GnuPG-bug-id: 5969