Wait for KeyCache in Readerstatus thread
When Kleopatra is not fully initialized the
updates from the ReaderStatus would block widgets
that try to react to the key infos received by
the Readerstatus.

This fixes a crash where a readerstatus update
with an uninitalized cache caused the openpgpwidget
to wait for the keycache even when it was destroyed
by closing the mainwindow.

part of:
GnuPG-Bug-Id: T5962