Ensure that the first child widget is visible when it gets focus
For all other child widgets this is already ensured by
QScrollArea::focusNextPrevChild(), but this method is not called if
the focus policy of the scroll area is set to NoFocus (which we often do
because it confuses many users if the scroll area gets focus instead of
some input field or button).

Reacting on any focus change is a bit crude, but anything else is much
more complicated or does not work, e.g. setting the first widget as focus
proxy of the scroll area or overwriting focusInEvent() to call
focusNextPrevChild().

GnuPG-bug-id: 5969