Make the automatic size adjustment work as expected
Adjusting the minimum height of the scroll area doesn't work correctly
because the main layout fails to calculate the correct minimal height
of the window. (It uses the minimumSizeHint of the info label which is
1 line of text although the info label shows 2 lines of text. This
causes the widget after the scroll area to be paint over the lower bit
of the scroll area.) Setting the minimum height of the scroll area also
had the disadvantage that it made it impossible for the user to reduce
the height of the window.

Instead we now simply resize the window as needed. Also rather crude, but
effective.

We limit the automatic size adjustment to 2/3 of the screen's size (i.e.
the same upper limit as QWidget::adjustSize()) instead of allowing the
user of ScrollArea to set an own limit for the height of the scroll area.

GnuPG-bug-id: 5969