Reduce the height of the minimum size hint of our scroll area
Do not add the height of the widget's minimum size hint to the scroll
area's minimum size hint. Adding the height made the minimum size of
the scroll area too large on small displays (or with large magnifaction)
and is not necessary because the scroll area enables vertical scrolling.
This reverts one of the changes made with d4d7bb80.

GnuPG-bug-id: 5969