Extract "Choose Protocol" dialog from NewCertificateWizard
This change converts the "Choose Protocol" page of the NewCertificateWizard
to a dialog of its own. This improves accessibility at least when using
orca because orca works very poorly with wizards, e.g. orca doesn't read
the title of the new page when the next page is shown in the wizard.

GnuPG-bug-id: 5832