Require C++20
We are well in 2022, so let's switch to C++20. I'm constantly trying
to write code that feels natural to me, only to discover "since C++20"
when I check why it doesn't work. And I'm not even talking about the
huge new features. Just those little improvements of the already existing
standard libraries.