Use separate dialogs instead of the wizard when creating OpenPGP keys
At least orca (screen reader) seems to have problems with wizards. If a
new page of the wizard is shown, then orca doesn't read the title/subtitle
of the new page, so that a user does really notice that a new page is
shown. Using separate dialogs for the different stages of key creation
avoids this problem.

GnuPG-bug-id: 5832