Allow setting a custom hint for name and email address
This allows organizations to provide a hint corresponding to the required
input pattern. Previously, this could be done with a placeholder text.
We use the placeholder text as fallback for the hint if no hint text is
configured. We do not set/use the placeholder text of the input fields
because placeholder texts are said to have bad accessibility (e.g. they
vanish as soon as one enters a character and they are not shown in the
first place if the input field already contains text).

The default hint "Must not include <, >, and @." for the name is removed.
Most people won't try to enter any of those characters as part of their
name and in case they do an appropriate error message is shown.

GnuPG-bug-id: 5832