Focus first child widget when switching the main view
This change adds a simple interface to allow passing input focus to a
suitable child widget of the different main views. The interface is
implemented by all main views.

The welcome view passes the focus to the label with the welcome message.
The certificate view passes the focus to the search bar.
The pad view passes the focus to the text edit, so that the user can
immediately paste or start typing.
The smart card view passes the focus to the first enabled child widget.

In particular, passing the focus to a child widget of the activated main
view removes the focus from the tool bar buttons that may have been used
to switch the main view.

GnuPG-bug-id: 6026