Make links in label text accessible
Add the AccessibleLink class (with role Link) that represents a link (or anchor)
of a rich text label. The anchors are enumerated and AccessibleRichTextLabel
returns them as accessible children. If a link gets the keyboard focus, then
HtmlLabel sends a corresponding focus event to the accessibility tools. This
makes links displayed by HtmlLabel accessible the same way as links displayed
by web browsers.

GnuPG-bug-id: 6034