Add button for showing the issuer certificate
Using a button for showing the issuer certificate instead of making this
functionality available via a link in the label text is much more
accessible. A link doesn't really tell the user what happens when it is
clicked. The button has a recognizable icon (same icon as the "Certificate
details" action) and a tool tip. Keyboard navigation is also clearer/
easier.

GnuPG-bug-id: 5843