Use a simple QWidget instead of a QScrollArea for the animated expander
Replacing the layout of the QScrollArea (with its internal widget/layout
hierarchy) messes up the widget hierarchy for accessibility tools, so
that screen readers (at least orca) do not read the texts of any widgets
in the content area of the animated expander. Using a simple QWidget
fixes this problem.

GnuPG-bug-id: 6046