Make the tool tip pop-ups in the Certify dialog accessible
At least on Linux, we need to post an ObjectShow event for the tool tip,
so that orca reads out the tool tip. Qt explicitly omits posting this
event because "Tooltips are read aloud twice in MS narrator".

Moreover, we change the time the tool tip will be displayed from
30 seconds to 24 hours.

GnuPG-bug-id: 6044