Remove informational text and add label for User ID table
For S/MIME, the informational text duplicates the "Related addresses"
label. And for OpenPGP, the text makes no sense for signing-only keys
and for keys without email addresses as user IDs. Therefore, we decided
to remove/hide it for S/MIME certificates and to use a simple "User IDs:"
label for OpenPGP keys. Setting the user ID table as buddy of the label
allows quick navigation to the table.

GnuPG-bug-id: 5843