On request, save decrypted file with embedded file name
If the "original" file name of the encrypted file is embedded in the
encrypted data and if it differs from the automatically chosen file name
derived from the encrypted file, then we ask the user whether the
decrypted file shall be saved with the "original" file name?
To minimize trouble for unattentive users, we do not use the embedded
file name if the default button is clicked.

Embedded file names containing backslashes are ignored to avoid trouble.
Any Unix-style paths are stripped from the embedded file names.

GnuPG-bug-id: 6056