Fix filtering of certificates by a text pattern
This is a really bad incompatibility between QRegExp and QRegularExpression
in QSortFilterProxyModel.
KeyListSortFilterProxyModel now uses filterRegularExpression() which
returns the regular expression set with setFilterRegularExpression().
Setting a filter pattern with setFilterFixedString() internally sets
a QRegExp, but not a QRegularExpression. Therefore, our filter model
used an empty QRegularExpression for filtering.
Setting the filter pattern with
setFilterRegularExpression(QRegularExpression::escape(...))
instead of setFilterFixedString(...) fixes this incompatibility.

GnuPG-bug-id: 6086