Improve cursor positioning of certificate input field
QLineEdit moves the cursor to the end when a text is set. This has the
effect that the start of a longer key/group summary is not visible when
we set this summary. This change improves the cursor positioning:
* Keep the cursor position when a key/group is updated in the background.
* Move the cursor to the start when the user finishes editing (with
Return), selects a key/group via the selection dialog, or moves focus to
another UI element.
* Otherwise, we keep the default behavior of QLineEdit.

GnuPG-bug-id: 6083