Ensure that current group is selected when list gets focus
By default, QListView (resp. QAbstractItemView) make the first visible
item the current item when the view gets focus, but it doesn't select
this item. OTOH, if one navigates the items with the keyboard, then
the current item is also selected. This change makes the behavior more
consistent and allows the user to interact immediately with the current
item after moving the focus to the list without having to select the
current item first.

GnuPG-bug-id: 6095