Add workaround for a bug in QListView::currentChanged
QListView::currentChanged sends an accessible focus event even if the
list view doesn't have focus. In this case the current item is saved
and restored when the key cache emits a changed signal. Restoring the
current item triggers the currentChanged signal.

The workaround skips the reimplementation of currentChanged in QListView
(which only sends the accessible focus event) by calling currentChanged
of QListView's parent class.

This fixes the problem that after opening the Edit Group dialog orca
thought that the current group in the Groups Configuration dialog
received the focus even though the Edit Group dialog has the input focus.

GnuPG-bug-id: 6095