Show a simple success message instead of a complicated result dialog
Instead of showing a result dialog with buttons for further actions we
decided to show a simple success message. All actions shown in the result
dialog are also available from the certificate list or the certificate
details view, and once the result dialog was closed the users anyway had
to resort to the actions in the menus.

GnuPG-bug-id: 5832