Simplify the revocation of certifications
Depending on what the user selected (key, one or more user IDs, a single
certification) all certifications that the user can revoke are
determined and, after confirmation, are revoked one after the other. If
a revocation is canceled (by canceling the pinentry dialog) or a
revocation fails then the operation is aborted.

GnuPG-bug-id: 6115