Make InfoField work as intended with Windows UI Automation
Representing the value widget of an InfoField as AccessibleValueLabel
with a custom role makes Windows UI Automation treat the value widget
as a custom widget providing a value instead of as static text. This
way the text of the label widget is used as accessible name and the
text of the value widget is used as accessible value as intended.

For Linux we keep the old behavior which works with orca/AT-SPI.

GnuPG-bug-id: 6149, 5843