Do not crash when closing details widget while certificate dump is shown
Closing the details widget triggers the destruction of its children. The
destruction of the DumpCertificateDialog triggers the finished signal
of the DumpCertificateCommand, but at this point ui.moreDetailsBtn has
already been destroyed.
The easiest solution is not to disable the moreDetailsBtn temporarily
since this allows us to remove the problematic connect. The "downside"
is that not disabling the button enables the user to open multiple
Certificate Dump dialogs, but that's not really a problem.

GnuPG-bug-id: 6180