Use GpgME::Error instead of the cryptic encoded error in task results
GpgME::Error provides a nice API, e.g. to check for a canceled operation,
while GpgME's encoded error (code) is a cryptic combination of the
error source and the actual error code which we would have to decode
when needed. And I didn't see a reason not to use GpgME::Error.

GnuPG-bug-id: 6196