Use error of decryption result also if decryption was canceled
This doesn't seem to change anything because logging shows that if the
decryption is canceled then the verification is also canceled (which
kind of makes sense because without prior decryption the verification
isn't possible). In any case, it feels more correct to return the
decryption error unless the decryption succeeded.

GnuPG-bug-id: 6196