Use UserIDs instead of indexes for identifying selected user IDs
This fixes a mix-up of user IDs caused by indexes calculated on all
user IDs vs. on valid user IDs only. Additionally, the improved
equality check for user IDs fixes a mix-up of user IDs with the same
text on the same key (e.g. if a user ID was revoked and then re-added).

GnuPG-bug-id: 6265