Clarify which smart card is meant in a question or message
If multiple smart cards are inserted, but only one smart card supports
the key algorithm, then Kleopatra doesn't ask which smart card the key
shall be written to. But for the user it may not be obvious which card
Kleopatra has chosen which makes it difficult or even dangerous to decide
whether it's okay to overwrite an existing key. To prevent mistakes we
now mention the card's display name instead of just "this card" in
messages/questions referring to a specific smart card.

GnuPG-bug-id: 5836