Avoid crashes when importing multiple files
Using a modal progress dialog caused reentrancy problems in importResult()
(because QProgressDialog::setValue() calls processEvents()) which lead to
two key list connections (in tryToFinish), subsequently two calls of
keyCacheUpdated() and eventually a crash. Since the imports happen in the
background in separate threads we do not need a modal dialog to have the
progress displayed to the user.

GnuPG-bug-id: 6183