Remove jobs from list after showing the error message
This fixes a possible crash: Removing the jobs before showing the error
message lead to further processing of the results while the error messages
were still shown which then lead to duplicate processing.

Moreover, we disconnect the job's connections to prevent duplicate calls,
e.g. if the command is canceled.

GnuPG-bug-id: 6183