Cache information from PIV card and store algo in key info
This makes the code for PIV cards more similar to the one for OpenPGP
cards. In particular, the algorithm of the card keys is now stored in
the key info instead of separately. Moreover, we cache the information
for the PIV card, so that we can always use it if we need to update
the displayed information.

GnuPG-bug-id: 6325