Only run selftest once on Windows
Running the selftest at each startup needlessly slows us down as
usually the problems occur either after installation / update or
never because GnuPG is part of the same Distribution as Kleopatra
is on Windows.

While under Linux it is more important to check for installation
incositencies. Under Windows it is also more rarely that
multiple versions of GnuPG run in the same home directory and
might interfer with their config files.

GnuPG-Bug-Id: T6346