Avoid asking multiple times whether to set owner trust
Imports of secret keys can result in multiple Imports for the same key
(once for the public key and once for the secret key). Don't ask the
users twice about the same keys. This will also avoid duplicate questions
if a public key file and a secret key file for the same key is imported.

GnuPG-bug-id: 6474