Avoid different dates for expiry of primary key and subkeys
Because the expiry of the primary key and the expiry of the subkeys have
to be changed with two calls there's a chance for a small difference
between the resulting expiration dates. To avoid the difference to cross
midnight (which results in a one-day difference in the UI) we give
Kleopatra (and the user if they need to enter a passphrase) a minute.

The real fix for this will happen in gpgme.

GnuPG-bug-id: 6473