Remove the output file if job was canceled or an error occurred
We make sure that an existing file is removed before the job is started.
If removing the existing file fails, then this is signalled by a "file
exists" error. Therefore, we don't try to remove the output file if this
error occurs. Otherwise, it should be safe to remove the output file.

GnuPG-bug-id: 6530