Use aboutToQuit signal to sync KSharedConfig
This seems more robust then relying on the destruction of
the KleopatraApplication object. I have tested it by logging
out as a Windows user from a Windows session. And received the
aboutToQuit but the dtor of the KleopatraApplication was not
called or at least at this point the sync did not work anymore.

GnuPG-Bug-Id: T6667