Only offer the user IDs to certify for certification
If the user wants to certify specific user IDs (selected in the
certificate details), then only offer those user IDs for certification.

And, when looking for existing tags, iterate only over those user IDs that
are available for certification instead of over all user IDs.
In particular, this also excludes expired and revoked user IDs when
looking for existing tags.

GnuPG-bug-id: 6697