Fix deselection/disabling of user ID of certification key
If one of the keys to certify is the last used certification key and
the tags need to be loaded, then sometimes (?) the user ID belonging to the
certification key wasn't deselected and disabled when opening the dialog.
This happened because the update of the selected user IDs didn't happen
after the user ID list was set up (after the tags were loaded).

This is fixed by updating the selected user IDs after setting up the
user ID list and after updating the selection based on the tags.

GnuPG-bug-id: 6469