Do not use applyWindowID for ViewEMailFilesDialog
When a dialog has a parent, while it is then centered on the
parent if that is visible does not get a Taskbar icon and is
better suited for Modal dialogs. Not applying the parent
window ID gives us the behavior that we are looking for that
it is shown grouped together in the windows Taskbar and the Window
can be selected and closed from there. For widgets this is different.