Add raise and activateWindow for mail dialog
Since it now does not have a parent and when the application
is hidden the dialog would not be raised properly. The
activate Window causes it to blink in the Taskbar on Windows
if it could not be raised. This is important as the dialog
can be triggered trough a filescommand and not only by a UI
action.