Abort group certification early on certain errors
If the user entered a wrong passphrase for the first certification or
ignored the pinentry until it timed out, then don't continue with the
next certificate, but abort the command. It's likely that the user wants
to restart the group certification instead of ending up with a partially
certified group.

GnuPG-bug-id: 6469