Always enable loading of remarks in the key cache
Remarks/tags are needed for some operations (e.g. certifications). For
other operations we need the certifications (e.g. to check on export if
there are uncertified keys). Enabling remarks in the key cache ensures
that the cache reloads the keys with certifications (signatures) and
remarks (signature notations) immediately after the initial fast key
listing.

GnuPG-bug-id: 6766