Change threading / blocking when loading aboutdata
When testing this I relalized that we might have races
where about data is read, e.g. in the WelcomeWidget and
where the custom about data is not yet loaded. So to
avoid having to update all places where about data is
used I would block this.

For me the backend Version check is also much more expensive
and that is only needed when the user clicks on
"Help about Kleopatra" so that could finish much later and is
put back in a thread.

(cherry picked from commit 09564b2613ac80834156c879bd15d2498a824b6d)