Add support for statusline and uidcomment
This adds support to put a custom statusline and a
uidcomment in the default openpgp key generation dialog.

I am aware that this can be bypassed by adding a user id
explicitly without such a uidcomment but it can also be bypassed
by using GnuPG on the command line. It is not really a
security feature, more of a slight marker that a key was
originally created with a version where we left a uidcomment
by default to indicate that the key was created with a version
that might have restricted support and should not be used
forever.

GnuPG-Bug-Id: T6797