Bump Kleopatra version to 3.2.0
Lots of changes since the last version bump and esp. the
addition of the mail viewer is a good argument for the
minor version bump. This also keeps it in line with the
upcoming Windows releases.

(cherry picked from commit 079836452b750b6703a46a324b0e33c1797c7a9e)