Show an error if the usage of key servers has been disabled
Instead of offering to use keys.gnupg.net as key server the commands are
aborted with an error. For very old versions of GnuPG (< 2.1.19) the
error message will be misleading but I think that's okay. In any case,
Kleopatra must not use keys.gnupg.net if the usage of key servers has
been disabled explicitly in GnuPG's configuration.

GnuPG-bug-id: 6761, 6866