Simplify lookup of key IDs prefixed with "0x"
It makes no sense to match strings of hex values that are optionally
prefixed with "0x" and then to check that the string doesn't start with
"0x". The WKD lookup is anyway only attempted if the query contains an
"@".

GnuPG-bug-id: 6868
(cherry picked from commit 249bfd0a8e812805cb782476fba8c4b69ab9c90f)