Only allow email queries if no key/directory servers are configured
If no key servers (OpenPGP) and no directory servers (S/MIME) are
configured then the server lookup only performs a WKD lookup. And for
WKD lookups we need an email address.

Add a hint that an email address has to be entered and validate that the
input looks (more or less) like an email address if only email queries
are possible.

GnuPG-bug-id: 6868