Remove confusing config dialog behavior from groups dialog
It was confusing that changes to the groups were not applied immediately.
This is fixed by making the groups dialog a simple QDialog instead of a
KConfigDialog with a config page. Remove the now superfluous mechanism
to set the correct initial focus. Simplify reloading the groups if they
were changed in the background. Add the Help button only if needed
(because adding and then hiding it breaks tab navigation).

The groups dialog is now managed by KleopatraApplication.

GnuPG-bug-id: 6662