Don't ask a second time for confirmation if a backup has been created
If the user has created a backup of the secret key and then chosen to
delete the copy on disk then don't annoy them with another request for
confirmation. Even if they accidentally chose to delete the copy on disk
they can restore it with the backup.

If the user hasn't created a backup (that we know of) then we keep
requesting confirmation.

GnuPG-bug-id: 6933