Fix update check for gpg4win
In gpg4win, Kleopatra has a version number like "3.2.0.231170 (Gpg4win-4.3.0)".
The regex extracting the gpg4win version number from this is not expecting a ')'
between the version number and the end of the string. This means that matching fails,
which causes the update check to be performed with an unexpected parameter.

GnuPG-Bug-ID: 6985
(cherry picked from commit 1ba561ebca006b7c2bec1b2d13389d931144debb)