Don't explicitely set a name for the first tab in the tab widget
If a tab doesn't have a name, the name of the selected key filter is used.
Only the first tab is given an explicit name by default, which is the same
as the namme of the default key filter ("All Certificates"). Since the name
was explicitely set, it will not update when changing the key filter in the
first tab, while it does update when it's not the first tab, creating a confusing
user experience.

Note that this change won't immediately fix the title for existing setups, since
the tab name is already stored in kleopatrarc.

GnuPG-Bug-ID: 7002