Fix when "Imported Certificates" tab is shown
Previously, there was some inconsistency, where the tab was shown when the key already existed, but not when it was actually a new key.
This behavior is now reversed. It also behaved differently depending on whether the key was imported using the button in kleopatra, by dragging
it into the window, or by clicking on it in the file manager.

GnuPG-Bug-ID: 6447
(cherry picked from commit 4aa9a854f78487891175781f7930c46350836f79)