Add error handling for Windows process connections
When trying to connect to an elevated responder window
the API calls to obtain the Process handle would fail.
This was unexpected and did not lead to an error but only to debug
messages. Now we show an error and exit to make the issue clear.

I also changed the variable names to make them more accurate.

GnuPG-Bug-Id: T7051
(cherry picked from commit 899e572e1097760972ecbf6ab2de3f5369096969)