Do not block application shutdown with a QEventLoopLocker
This isn't necessary (anymore) and it prevents Kleopatra from shutting
down when quitOnLastWindowClosed is enabled (which we want to do for users
with elevated permissions on Windows).

Using a QEventLoopLocker and setting quitOnLastWindowClosed both prevent
the application from quitting when the last window is closed. For
Kleopatra it's sufficient to use the latter method.

GnuPG-bug-id: 7050
(cherry picked from commit bfed041ea1b1524de14214ee70f73bc9b3e23a25)