Remove showToolTip helper
It doesn't build on Windows anymore because QDesktopWidget is gone in Qt 6.
Qt 6 doesn't assign a parent to the tool tip label anymore, so that the
current code cannot be fixed easily for Windows. The only remaining usage
of this helper was removed with the previous commit.

GnuPG-bug-id: 7073
(cherry picked from commit be87086732d4b629ffcdb4b6847869b0813a2625)