CertificateDetailsDialog: Do not parent to main window
Currently, Kleopatra's dialogs behave in an unexpected way: While a dialog is open, the user can still interact with the main window.
The dialog cannot move behind the main window, however. It makes sense to make dialogs either entirely modal, or not modal at all.
For this dialog, since it can be opened multiple times simultaneously (including from itself), it makes sense to have it non-modal,
allowing aribitrary moving of the dialogs and the main window on top of each other. This is done by removing the parenting to the main window
(or a different dialog).

GnuPG-Bug-ID: 7094