Add debug view
Add a dialog that shows the output of certain predefined commands, which can be selected from a combobox.
This can be useful for support scenarios.
The user is given the options to save this data to a file or copy it to the clipboard.

GnuPG-Bug-ID: 6072
(cherry picked from commit 9b676c51f539ce8356621e36981ace2848881e20)