Use KConfig on Windows for KMessageBox don't ask me again checkboxes
This is an exact copy of the KConfig storage backend for the KMessageBox
in framework integration, which is usually only available on Plasma.
This ensure the don't ask me again state is stored in the same location
as the rest of the KConfig powered configs in kleopatrarc also on
Windows.

GnuPG-bug-Id: T6799