Include keygrip when updating key
When clicking on a certificate in the list, the details dialog is opened and a key listing for this certificate is started.
This job did previously not ask for the keygrip to be included in the result. This means that the keycache is updated to include
the key without keygrip information, which results in the key list being updated with the new key, causing the content of the keygrip column
for this certificate to disappear until kleopatra is restarted or something triggers an update to the entire key list.

(cherry picked from commit afdd195917a28e42e4218095a246ef1a4e41ff27)
GnuPG-bug-id: 6957