Fix glitches in certificate list
When opening the sign/encrypt or decrypt/verify dialog, the certificate list in the main window stops scrolling correctly;
the scrollbar continues moving, but the table's content does not move accordingly. (This seems to be specific to linux or even wayland)

This is caused by {DecryptVerfyFilesDialog, SignEncryptFilesWizard}::readConfig calling QWidget::winId, which can cause unexpected behavior and is unnecessary, since we can resize the window in a much simpler way.